+++ src/parts/Content.tsx
@@ -1,5 +1,5 @@
 import DateFnsUtils from "@date-io/date-fns";
-import { AttributeData, ControlsValue, Screen, Step } from "@decisively-io/interview-sdk";
+import type { AttributeData, ControlsValue, Screen, Step } from "@decisively-io/interview-sdk";
 import { CircularProgress } from "@material-ui/core";
 import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
@@ -7,10 +7,10 @@ import { MuiPickersUtilsProvider } from "@material-ui/pickers";
 import React, { useContext, useImperativeHandle } from "react";
 import { FormProvider, useForm } from "react-hook-form";
 import styled from "styled-components";
-import { DISPLAY_NAME_PREFIX } from "../Constants";
+import { DISPLAY_NAME_PREFIX, LOADING_ANIMATION_CSS } from "../Constants";
 import { generateValidator } from "../util/Validation";
 import { deriveDefaultControlsValue } from "../util/controls";
-import Controls, { ControlComponents } from "./controls";
+import Controls, { type ControlComponents } from "./controls";
 import { InterviewContext } from "./index";
 
 export const classes = {
@@ -87,6 +87,8 @@ const Wrap = styled.form`
 `;
 
 export const StyledControlsWrap = styled.div`
+  ${LOADING_ANIMATION_CSS}
+
   > *:not(:last-child) {
     margin-bottom: 1rem;
   }