+++ src/parts/controls/ImageControlWidget.tsx
@@ -1,7 +1,7 @@
-import { ImageControl } from "@decisively-io/interview-sdk";
+import type { ImageControl } from "@decisively-io/interview-sdk";
 import React from "react";
 import { DISPLAY_NAME_PREFIX } from "./ControlConstants";
-import { ControlWidgetProps } from "./ControlWidgetTypes";
+import type { ControlWidgetProps } from "./ControlWidgetTypes";
 import FormControl from "./FormControl";
 
 export interface ImageControlWidgetProps extends ControlWidgetProps<ImageControl> {}
@@ -19,13 +19,13 @@ export const ImageControlWidget = Object.assign(
   }),
   {
     displayName: `${DISPLAY_NAME_PREFIX}/Image`,
-    /*** @deprecated use `ImageInput` directly */
+    /*** @deprecated use `ImageControlWidget` directly */
     _: null as any as React.ComponentType<ImageControlWidgetProps>,
   },
 );
 ImageControlWidget._ = ImageControlWidget;
 
-/*** @deprecated use `ImageInput` directly */
+/*** @deprecated use `ImageControlWidget` directly */
 export const _ = ImageControlWidget;
 
 export default ImageControlWidget;