+++ src/parts/controls/TypographyControlWidget.tsx
@@ -1,11 +1,12 @@
-import { TypographyControl } from "@decisively-io/interview-sdk";
-import MuiTypography, { TypographyProps as MuiTypographyProps } from "@material-ui/core/Typography";
+import type { TypographyControl } from "@decisively-io/interview-sdk";
+import MuiTypography, { type TypographyProps as MuiTypographyProps } from "@material-ui/core/Typography";
 import React from "react";
 import styled from "styled-components";
+import { LOADING_ANIMATION_CSS } from "../../Constants";
 import { DISPLAY_NAME_PREFIX } from "./ControlConstants";
-import { ControlWidgetProps } from "./ControlWidgetTypes";
+import type { ControlWidgetProps } from "./ControlWidgetTypes";
 
-export interface TypographyControlWidgetProps extends ControlWidgetProps<TypographyControl> {
+export interface TypographyControlWidgetProps extends ControlWidgetProps<TypographyControl & { loading?: boolean }> {
   typographyProps?: MuiTypographyProps;
   className?: string;
 }
@@ -36,16 +37,26 @@ const BannerYellow = styled(BannerBase)`
   border-left: 4px solid rgb(252, 180, 20);
   background-color: rgb(255, 249, 236);
 `;
+const StyledTypography = styled(MuiTypography)`
+  ${LOADING_ANIMATION_CSS}
+`;
 
-type BannerStyle = Extract<TypographyControl["style"], "banner-yellow" | "banner-green" | "banner-red">;
+interface BannerProps {
+  control: TypographyControl & { loading?: boolean };
+  text: string;
+}
 
-const Banner: React.FC<{ text: string; style: BannerStyle; emoji?: TypographyControl["emoji"] }> = React.memo(({ style, text, emoji }) => {
+const Banner = React.memo((props: BannerProps) => {
+  const { control, text } = props;
+  const { style, emoji } = control;
   const Comp = style === "banner-green" ? BannerGreen : style === "banner-red" ? BannerRed : BannerYellow;
 
   return (
     <Comp>
       <span className="emoji" dangerouslySetInnerHTML={{ __html: emoji || "" }} />
-      <MuiTypography variant="body1">{text}</MuiTypography>
+      <StyledTypography data-loading={control.loading} variant="body1">
+        {text}
+      </StyledTypography>
     </Comp>
   );
 });
@@ -61,19 +72,11 @@ export const TypographyControlWidget = Object.assign(
       <div className={className}>
         {(() => {
           if (style === "banner-red" || style === "banner-green" || style === "banner-yellow") {
-            return (
-              <Banner
-                {...{
-                  style,
-                  text,
-                  emoji,
-                }}
-              />
-            );
+            return <Banner control={control} text={text} />;
           }
 
           return (
-            <MuiTypography variant={style} {...typographyProps}>
+            <StyledTypography data-loading={control.loading} variant={style} {...typographyProps}>
               {emoji === undefined ? null : (
                 <span className="emoji">
                   {emoji}
@@ -82,7 +85,7 @@ export const TypographyControlWidget = Object.assign(
               )}
 
               <span className="text">{text}</span>
-            </MuiTypography>
+            </StyledTypography>
           );
         })()}
       </div>
@@ -90,13 +93,13 @@ export const TypographyControlWidget = Object.assign(
   }),
   {
     displayName: DISPLAY_NAME,
-    /*** @deprecated use `Typography` directly instead */
+    /*** @deprecated use `TypographyControlWidget` directly instead */
     _: null as any as React.ComponentType<TypographyControlWidgetProps>,
   },
 );
 TypographyControlWidget._ = TypographyControlWidget;
 
-/*** @deprecated use `Typography` directly instead */
+/*** @deprecated use `TypographyControlWidget` directly instead */
 export const _ = TypographyControlWidget;
 
 export default TypographyControlWidget;