+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+## 0.3.0
+
+- **BREAKING** pass in session instance directly
+- add support for dynamic containers
+- deprecate `Parts`, exports them directly
+
 ## 0.2.0
 
 - rename all control components to `XControlWidget` instead of just `X` to avoid colliding with inbuilt names such as `Date`
 - use interview-sdk's re-export of interview types
-- add support for conditional containers